requirements.txt
beautifulsoup4~=4.13.0b2
contexttimer~=0.3.3
gspread~=5.8.0
lxml~=4.9.2
pyperclip~=1.9.0
python-dateutil~=2.9.0.post0
pytubefix~=6.5.1
requests~=2.32.3
selenium~=4.23.1
scrapetube~=2.5.1
scrython~=1.11.0
tqdm~=4.66.4
Unidecode~=1.3.8
youtube-search-python~=1.6.6
